feat: releases and changelog
Add configuration for release and changelog management based on
[release-please](https://github.com/googleapis/release-please) and
[Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/#summary).

Release-please generates a release PR that's updated on every merge
to `main`. Once merged, it generates changelog, updates version,
and creates a tag and a GitHub release. The process is mostly
automatic but requires that commit follow the Conventional Commits
standard which is enforced by adding an action that lints PR titles
for following the standard (we use squash, so the tile is the most
important part, convetional commits should be followed for commits
to though).